рендер карточек в доп блоки
