+++ src/view/extra-list.js
@@ -1,29 +1,39 @@
 import {createElement} from "../utils.js";
+import {renderFilmCard} from "./render-film-card.js";
 
-const createExtraListTemplate = (cardsTemplate, title) => {
+const createExtraListTemplate = (title) => {
   return (`<section class="films-list--extra">
             <h2 class="films-list__title">${title}</h2>
             <div class="films-list__container">
-              ${cardsTemplate}
+
             </div>
           </section>`);
 };
 
 export default class ExtraList {
-  constructor(cardsTemplate, blockTiltle) {
+  constructor(blockTiltle, childsData) {
     this._element = null;
     this._blockTiltle = blockTiltle;
-    this._cardsTemplate = cardsTemplate;
+    this._childsData = childsData;
   }
 
   _getTemplate() {
-    return createExtraListTemplate(this._cardsTemplate, this._blockTiltle);
+    return createExtraListTemplate(this._blockTiltle);
   }
 
   getElement() {
     if (!this._element) {
       this._element = createElement(this._getTemplate());
     }
+    return this._element;
+  }
+
+  getElementWithChildren() {
+    const filmList = this.getElement().querySelector(`.films-list__container`);
+    this._childsData.forEach((element) => {
+      renderFilmCard(filmList, element);
+    });
+
     return this._element;
   }
